+++ Server/Models/Monsters/ChristmasMonster.cs
@@ -24,7 +24,7 @@ namespace Server.Models.Monsters
             base.RefreshStats();
         }
 
-        public override int Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override long Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
             return base.Attacked(attacker, 1, element, canReflect, ignoreShield, false, canStruck);
         }
